PostgresNodePortManager(+company) was moved in own file - port_manager.py
